feat(client): exposed method for making raw function requests
